requirements.txt
Flask
Werkzeug
pandas
scikit-learn
numpy
textblob
pyspellchecker